+++ pdfimpose/__init__.py
@@ -140,7 +140,7 @@ import PyPDF2
 
 from pdfimpose import errors
 
-VERSION = "0.1.1"
+VERSION = "1.0.0"
 __AUTHOR__ = "Louis Paternault (spalax+python@gresille.org)"
 __COPYRIGHT__ = "(C) 2014-2017 Louis Paternault. GNU GPL 3 or later."